pixelator_core_py

Function run_leiden

pub fn run_leiden(
    parquet_file: &str,
    resolution: f64,
    max_iteration: Option<usize>,
    partition: Option<Vec<usize>>,
    output: &str,
    randomness: f64,
    seed: Option<u64>,
    merge_edge_threshold: Option<usize>,
    merge_edge_threshold_relative: Option<f64>,
) -> PyResult<(usize, f64)>

Finds community partitioning using the Leiden algorithm

Arguments

parquet_file - Path to the Parquet file containing the edge list.

max_iteration - maximum number of iterations to perform. NB: in most cases the algorithm should converge and stop by itself.

partition - initial node partitioning to use. If not provided, singlet partitions will be used

resolution - resolution to use in the quality function. The larger the resolution, the smaller the resulting communities will be.

output - Path to the output parquet file. Default is node_partitions.parquet.

randomness - Randomness of node transitions. Low values will favor moves maximizing the quality, while higher values will allow suboptimal moves, making it easier to avoid local minima at the cost of convergence speed. Default value is 0.1

seed - seed to use in the random generator

merge_edge_threshold - Merge communities when their connecting edge count exceeds this absolute threshold. Cannot be used together with merge_edge_threshold_relative.

merge_edge_threshold_relative - Merge communities when their connecting edge relative to the number of nodes in the smallest community exceeds this threshold. Cannot be used together with merge_edge_threshold.

Returns

A tuple containing:

The number of partitions

The overall quality of the partitioning

The node partitioning is written in a Parquet file as specified by the output parameter.